Remove non-functional AI poisoning
Several models seem to blatantly ignore the robots.txt. This number
dropped after a 0-byte favicon was added (not sure why, but it seems a
couple models interpreted the 404 from the favicon as a 404 from the
robots.txt?), but is still non-zero.

Duck.ai and OpenAI both blatantly ignore the robots.txt

The AI poisoning did appear to work at first, but it stopped working
shortly after. Keeping it in the footer was likely inefficient. Certain
queries result in the AI note being mentioned, but not respected, which
isn't much of a surprise. There's significant amounts of preprocessing
involved to stop precisely what I wanted to do here from working. It
might've worked with a much longer injection query, but I do not feel
like adding that.

I'm pretty sure the conclusion is just "live with it", seeing as
@openai, @duckduckgo, and probably many others simply cannot be bothered
respecting robots.txt.
I did expect this from OpenAI (seeing as they love pirating books:
https://www.theatlantic.com/technology/archive/2025/03/libgen-meta-openai/682093/),
but not DDG.

This is where I would go full nuclear and block their IP ranges
associated with AI use, as well as doing low-level filtering, but
because github pages, I obviously don't have access to do that, and I'm
not going to self-host a public-facing website. Guess I'll just have to
live with AI slurping up all the data.
In retrospect, it doesn't really matter anyway. The code is hosted on
GitHub, and all of GitHub is almost certainly being thrown straight at
Microsoft and OpenAI's training systems.

It really is just dystopian